Improved codebase organization
