--- frontend/lib/Soporte.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-
-// class SoporteMegafonitoScreen extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       appBar: AppBar(
-//         title: Text(
-//           'Soporte Megafonito',
-//           style: TextStyle(color: Colors.white),
-//         ),
-//         backgroundColor: Color(0xFF14213D),
-//       ),
-//       backgroundColor: Colors.white, // Correctly placed here
-//       body: Center(
-//         child: Text('Aquí va la sección de soporte.'),
-//       ),
-//     );
-//   }
-// }
-
-class SoporteMegafonitoScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.white,
-      child: Center(
-        child: Text('Aquí va la sección de soporte.'),
-      ),
-    );
-  }
-}
