+++ frontend/lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-import 'Anuncios.dart';
-import 'service.dart'; // Update path to match your structure
+import 'screens/splash_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,7 +10,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      title: 'Login Demo',
+      title: 'Megafonito',
       theme: ThemeData(
         appBarTheme: AppBarTheme(
           iconTheme: IconThemeData(color: Colors.white),
@@ -25,472 +23,6 @@ class MyApp extends StatelessWidget {
         ),
       ),
       home: SplashScreen(),
-    );
-  }
-}
-
-class SplashScreen extends StatefulWidget {
-  @override
-  _SplashScreenState createState() => _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen> {
-  double progress = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    _startLoading();
-  }
-
-  void _startLoading() {
-    Timer.periodic(Duration(milliseconds: 100), (timer) {
-      setState(() {
-        progress += 0.1;
-      });
-
-      if (progress >= 1) {
-        timer.cancel();
-        Navigator.of(context).pushReplacement(
-          MaterialPageRoute(builder: (context) => LoginScreen()),
-        );
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xFF14213D),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset(
-              'assets/LogoMegafon_V2.png',
-              width: 200,
-              height: 200,
-            ),
-            SizedBox(height: 20),
-            Text(
-              'Cargando...',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: Color(0xFFFCA311),
-              ),
-            ),
-            SizedBox(height: 20),
-            Container(
-              width: 300,
-              child: LinearProgressIndicator(
-                value: progress,
-                minHeight: 20,
-                backgroundColor: Color(0xFFE5E5E5),
-                valueColor: AlwaysStoppedAnimation<Color>(Color(0xFFFCA311)),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class User {
-  String name;
-  String email;
-  String password;
-
-  User({required this.name, required this.email, required this.password});
-}
-
-class LoginScreen extends StatefulWidget {
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  final _formKey = GlobalKey<FormState>();
-  String? username;
-  String? password;
-  bool _isLoading = false;
-// Remove the hardcoded user list since we'll get this from backend
-
-  void _login() async {
-    if (_formKey.currentState!.validate()) {
-      setState(() {
-        _isLoading = true;
-      });
-
-      try {
-        // Call backend API to login
-        final result = await ApiService.login(username!);
-
-        // Successfully logged in
-        setState(() {
-          _isLoading = false;
-        });
-
-        // Navigate to AnunciosScreen with the user role info
-        Navigator.of(context).pushReplacement(
-          MaterialPageRoute(
-            builder: (context) => AnunciosScreen(
-              userName: username!, // Added user name
-              userEmail: username!, // Using username as email for now
-              isSuperUser: ApiService.userRole == 'admin',
-            ),
-          ),
-        );
-      } catch (e) {
-        setState(() {
-          _isLoading = false;
-        });
-
-        // Show error message
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Error de inicio de sesión: $e')),
-        );
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Image.asset(
-                'assets/LogoMegafon_V2.png',
-                width: 150,
-                height: 150,
-              ),
-              SizedBox(height: 20),
-              Card(
-                elevation: 10,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20),
-                ),
-                child: Container(
-                  padding: const EdgeInsets.all(24.0),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(20),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black26,
-                        blurRadius: 8.0,
-                        spreadRadius: 2.0,
-                      ),
-                    ],
-                  ),
-                  child: Form(
-                    key: _formKey,
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        Text(
-                          'Iniciar Sesión',
-                          style: TextStyle(
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                            color: Color(0xFF14213D),
-                          ),
-                        ),
-                        SizedBox(height: 20),
-                        TextFormField(
-                          decoration: InputDecoration(
-// labelText: 'Correo Electrónico',
-                            labelText: 'Nombre de Usuario',
-                            labelStyle: TextStyle(color: Color(0xFF14213D)),
-                            filled: true,
-                            fillColor: Color(0xFFE5E5E5),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(8.0),
-                              borderSide: BorderSide(color: Color(0xFF14213D)),
-                            ),
-                          ),
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return 'Por favor ingresa tu nombre de usuario';
-                            }
-                            return null;
-                          },
-                          onChanged: (value) {
-                            username = value;
-                          },
-                        ),
-                        SizedBox(height: 20),
-                        TextFormField(
-                          decoration: InputDecoration(
-                            labelText: 'Contraseña',
-                            labelStyle: TextStyle(color: Color(0xFF14213D)),
-                            filled: true,
-                            fillColor: Color(0xFFE5E5E5),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(8.0),
-                              borderSide: BorderSide(color: Color(0xFF14213D)),
-                            ),
-                          ),
-                          obscureText: true,
-                          // validator: (value) {
-                          //   if (value == null || value.isEmpty) {
-                          //     return 'Por favor ingresa tu contraseña';
-                          //   }
-                          //   return null;
-                          // },
-                          onChanged: (value) {
-                            password = value;
-                          },
-                        ),
-                        SizedBox(height: 20),
-                        _isLoading
-                            ? CircularProgressIndicator(
-                                color: Color(0xFFFCA311))
-                            : ElevatedButton(
-                                onPressed: _login,
-                                style: ElevatedButton.styleFrom(
-                                  backgroundColor: Color(0xFFFCA311),
-                                  padding: EdgeInsets.symmetric(
-                                      vertical: 12, horizontal: 20),
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                ),
-                                child: Text(
-                                  'Iniciar Sesión',
-                                  style: TextStyle(fontSize: 16),
-                                ),
-                              ),
-                        SizedBox(height: 20),
-                        TextButton(
-                          onPressed: () {
-                            // Password recovery action
-                          },
-                          child: Text(
-                            '¿Olvidaste tu contraseña?',
-                            style: TextStyle(color: Color(0xFF14213D)),
-                          ),
-                        ),
-                        TextButton(
-                          onPressed: () {
-                            Navigator.of(context).push(
-                              MaterialPageRoute(
-                                builder: (context) => RegisterScreen(),
-                              ),
-                            );
-                          },
-                          child: Text(
-                            '¿Quieres registrarte?',
-                            style: TextStyle(color: Color(0xFF14213D)),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class RegisterScreen extends StatefulWidget {
-  @override
-  _RegisterScreenState createState() => _RegisterScreenState();
-}
-
-class _RegisterScreenState extends State<RegisterScreen> {
-  final _formKey = GlobalKey<FormState>();
-  String? name;
-  String? email;
-  String? password;
-  bool _isLoading = false;
-
-  void _register() async {
-    if (_formKey.currentState!.validate()) {
-      setState(() {
-        _isLoading = true;
-      });
-
-      try {
-        // Here you would call your API service to register the user
-        // Example: await ApiService.register(name!, email!, password!);
-
-        // For now we'll just simulate success
-        await Future.delayed(Duration(seconds: 1));
-        setState(() {
-          _isLoading = false;
-        });
-
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Usuario registrado con éxito')),
-        );
-        Navigator.of(context).pop(); // Return to login screen
-      } catch (e) {
-        setState(() {
-          _isLoading = false;
-        });
-
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Error al registrar usuario: $e')),
-        );
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Registrar Usuario'),
-        backgroundColor: Color(0xFF14213D),
-      ),
-      body: Center(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Image.asset(
-                'assets/LogoMegafon_V2.png',
-                width: 150,
-                height: 150,
-              ),
-              SizedBox(height: 20),
-              Card(
-                elevation: 10,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20),
-                ),
-                child: Container(
-                  padding: const EdgeInsets.all(24.0),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(20),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black26,
-                        blurRadius: 8.0,
-                        spreadRadius: 2.0,
-                      ),
-                    ],
-                  ),
-                  child: Form(
-                    key: _formKey,
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        Text(
-                          'Registrar Usuario',
-                          style: TextStyle(
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                            color: Color(0xFF14213D),
-                          ),
-                        ),
-                        SizedBox(height: 20),
-                        TextFormField(
-                          decoration: InputDecoration(
-                            labelText: 'Nombre',
-                            labelStyle: TextStyle(color: Color(0xFF14213D)),
-                            filled: true,
-                            fillColor: Color(0xFFE5E5E5),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(8.0),
-                              borderSide: BorderSide(color: Color(0xFF14213D)),
-                            ),
-                          ),
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return 'Por favor ingresa tu nombre';
-                            }
-                            return null;
-                          },
-                          onChanged: (value) {
-                            name = value;
-                          },
-                        ),
-                        SizedBox(height: 20),
-                        TextFormField(
-                          decoration: InputDecoration(
-                            labelText: 'Correo Electrónico',
-                            labelStyle: TextStyle(color: Color(0xFF14213D)),
-                            filled: true,
-                            fillColor: Color(0xFFE5E5E5),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(8.0),
-                              borderSide: BorderSide(color: Color(0xFF14213D)),
-                            ),
-                          ),
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return 'Por favor ingresa tu correo electrónico';
-                            }
-                            return null;
-                          },
-                          onChanged: (value) {
-                            email = value;
-                          },
-                        ),
-                        SizedBox(height: 20),
-                        TextFormField(
-                          decoration: InputDecoration(
-                            labelText: 'Contraseña',
-                            labelStyle: TextStyle(color: Color(0xFF14213D)),
-                            filled: true,
-                            fillColor: Color(0xFFE5E5E5),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(8.0),
-                              borderSide: BorderSide(color: Color(0xFF14213D)),
-                            ),
-                          ),
-                          obscureText: true,
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return 'Por favor ingresa tu contraseña';
-                            }
-                            return null;
-                          },
-                          onChanged: (value) {
-                            password = value;
-                          },
-                        ),
-                        SizedBox(height: 20),
-                        _isLoading
-                            ? CircularProgressIndicator(
-                                color: Color(0xFFFCA311))
-                            : ElevatedButton(
-                                onPressed: _register,
-                                style: ElevatedButton.styleFrom(
-                                  backgroundColor: Color(0xFFFCA311),
-                                  padding: EdgeInsets.symmetric(
-                                      vertical: 12, horizontal: 20),
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                ),
-                                child: Text(
-                                  'Registrar',
-                                  style: TextStyle(fontSize: 16),
-                                ),
-                              ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
