+++ frontend/lib/screens/login_screen.dart
@@ -1,0 +1,190 @@
+// lib/screens/login_screen.dart
+import 'package:flutter/material.dart';
+import '../services/auth_service.dart';
+import 'anuncios_screen.dart';
+import 'register_screen.dart';
+
+class LoginScreen extends StatefulWidget {
+  @override
+  _LoginScreenState createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  final _formKey = GlobalKey<FormState>();
+  String? username;
+  String? password;
+  bool _isLoading = false;
+
+  void _login() async {
+    if (_formKey.currentState!.validate()) {
+      setState(() {
+        _isLoading = true;
+      });
+
+      try {
+        // Call auth service to login
+        final user = await AuthService.login(context, username!);
+
+        setState(() {
+          _isLoading = false;
+        });
+
+        // Navigate to AnunciosScreen with the user role info
+        Navigator.of(context).pushReplacement(
+          MaterialPageRoute(
+            builder: (context) => AnunciosScreen(
+              userName: user.name,
+              // userName: widget.userName,
+              userEmail: user.email,
+              isSuperUser: AuthService.isAdmin,
+            ),
+          ),
+        );
+      } catch (e) {
+        setState(() {
+          _isLoading = false;
+        });
+        // Error is already shown by the service
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: SingleChildScrollView(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Image.asset(
+                'assets/LogoMegafon_V2.png',
+                width: 150,
+                height: 150,
+              ),
+              SizedBox(height: 20),
+              Card(
+                elevation: 10,
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(20),
+                ),
+                child: Container(
+                  padding: const EdgeInsets.all(24.0),
+                  decoration: BoxDecoration(
+                    color: Colors.white,
+                    borderRadius: BorderRadius.circular(20),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.black26,
+                        blurRadius: 8.0,
+                        spreadRadius: 2.0,
+                      ),
+                    ],
+                  ),
+                  child: Form(
+                    key: _formKey,
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Text(
+                          'Iniciar Sesión',
+                          style: TextStyle(
+                            fontSize: 24,
+                            fontWeight: FontWeight.bold,
+                            color: Color(0xFF14213D),
+                          ),
+                        ),
+                        SizedBox(height: 20),
+                        TextFormField(
+                          decoration: InputDecoration(
+                            labelText: 'Nombre de Usuario',
+                            labelStyle: TextStyle(color: Color(0xFF14213D)),
+                            filled: true,
+                            fillColor: Color(0xFFE5E5E5),
+                            border: OutlineInputBorder(
+                              borderRadius: BorderRadius.circular(8.0),
+                              borderSide: BorderSide(color: Color(0xFF14213D)),
+                            ),
+                          ),
+                          validator: (value) {
+                            if (value == null || value.isEmpty) {
+                              return 'Por favor ingresa tu nombre de usuario';
+                            }
+                            return null;
+                          },
+                          onChanged: (value) {
+                            username = value;
+                          },
+                        ),
+                        SizedBox(height: 20),
+                        TextFormField(
+                          decoration: InputDecoration(
+                            labelText: 'Contraseña',
+                            labelStyle: TextStyle(color: Color(0xFF14213D)),
+                            filled: true,
+                            fillColor: Color(0xFFE5E5E5),
+                            border: OutlineInputBorder(
+                              borderRadius: BorderRadius.circular(8.0),
+                              borderSide: BorderSide(color: Color(0xFF14213D)),
+                            ),
+                          ),
+                          obscureText: true,
+                          onChanged: (value) {
+                            password = value;
+                          },
+                        ),
+                        SizedBox(height: 20),
+                        _isLoading
+                            ? CircularProgressIndicator(
+                                color: Color(0xFFFCA311))
+                            : ElevatedButton(
+                                onPressed: _login,
+                                style: ElevatedButton.styleFrom(
+                                  backgroundColor: Color(0xFFFCA311),
+                                  padding: EdgeInsets.symmetric(
+                                      vertical: 12, horizontal: 20),
+                                  shape: RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                ),
+                                child: Text(
+                                  'Iniciar Sesión',
+                                  style: TextStyle(fontSize: 16),
+                                ),
+                              ),
+                        SizedBox(height: 20),
+                        TextButton(
+                          onPressed: () {
+                            // Password recovery action
+                          },
+                          child: Text(
+                            '¿Olvidaste tu contraseña?',
+                            style: TextStyle(color: Color(0xFF14213D)),
+                          ),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).push(
+                              MaterialPageRoute(
+                                builder: (context) => RegisterScreen(),
+                              ),
+                            );
+                          },
+                          child: Text(
+                            '¿Quieres registrarte?',
+                            style: TextStyle(color: Color(0xFF14213D)),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
